analyze services to test services
